requirements.txt
streamlit
tensorflow
numpy
Pillow
matplotlib
seaborn
opencv-python